requirements.txt
streamlit
PyMuPDF
python-dotenv
langchain
langchain-google-genai
faiss-cpu